+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - (Bugfix) (Platform) Fix Container Resource Adjustments
 - (Bugfix) (Platform) Fix LM CLI Option
 - (Bugfix) (Platform) Fix topology for Gateways
+- (Feature) (Platform) Dump CLI switch to Services
 
 ## [1.3.2](https://github.com/arangodb/kube-arangodb/tree/1.3.2) (2025-11-20)
 - (Bugfix) (Platform) Increase memory limit for Inventory